Allow proper Getter & Setter entity properties
They allow you to pass custom parameters to your getters/setters. This
way, you can pass the pointer to the component in a getter/setter that
is not located in the component, but is for example managed by the
receipt itself.

Unfortunately, thes getters introduce a lot of return-copying. Although,
properties are not meant for super high speed r/w we'll definitely have
to test those against performance.