Implement MUCH more efficient terrain collision detection
Previously, we used a general simple quadtree. However, building this thing was
expensive, while the vertices and triangles in the terrain mesh were actually
ordered in memory, so direct access is possible. This new method exploits this
order to avoid building a huge tree and to do collision detection in O(1)